Promote DurationString, RawRequest, RawResponse out of internal/
These three types were already part of the public API (re-exported from
encore.dev/api and encore.dev/pubsub), but their definitions lived in
internal/ which TypeDoc excludes. Result: users saw the type names in
public signatures with no documentation page to follow.

Moves:
- internal/types/mod.ts -> types/mod.ts (just DurationString; the file
  contained nothing else, so the internal/types directory is removed)
- internal/api/node_http.ts -> api/node_http.ts (RawRequest, RawResponse,
  and their private DummySocket helper)

Imports updated in cron/mod.ts, pubsub/subscription.ts, pubsub/mod.ts,
api/mod.ts, and internal/appinit/mod.ts. All three types now have full
doc pages.

Adds 'durationComponent' to intentionallyNotExported (private type-level
machinery for DurationString, same pattern as brandedAttribute) and a
node.js externalSymbolLinkMappings entry for the {@link pipe} reference
inherited from @types/node's stream.Writable JSDoc.
