+++ runtimes/js/encore.dev/internal/appinit/mod.ts
@@ -2,7 +2,7 @@ import { isMainThread, Worker, workerData } from "node:worker_threads";
 import { Gateway } from "../../api/gateway";
 import { Middleware, MiddlewareRequest, HandlerResponse } from "../../api/mod";
 import { IterableSocket, IterableStream, Sink } from "../../api/stream";
-import { RawRequest, RawResponse } from "../api/node_http";
+import { RawRequest, RawResponse } from "../../api/node_http";
 import { setCurrentRequest } from "../reqtrack/mod";
 import * as runtime from "../runtime/mod";
 import { fileURLToPath } from "node:url";
